Plivo New Updated changes
Commit 2

+++ src/com/agilecrm/core/api/widgets/WidgetsAPI.java
@@ -313,6 +313,8 @@ public class WidgetsAPI
 	{
 
 		Widget widget = WidgetUtil.getWidget(id);
-		widget.delete();
+		if (widget != null)
+			widget.delete();
+		System.err.println("The widget is null and id is " + id);
 	}
 }
